+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/__init__.py
@@ -9,7 +9,7 @@ from ._form_recognizer_client import FormRecognizerClient
 from ._form_training_client import FormTrainingClient
 
 from ._models import (
-    FormContent,
+    FormElement,
     LengthUnit,
     TrainingStatus,
     CustomFormModelStatus,
@@ -41,7 +41,7 @@ __all__ = [
     'TrainingStatus',
     'CustomFormModelStatus',
     'FormContentType',
-    'FormContent',
+    'FormElement',
     'FormTable',
     'FormTableCell',
     'TrainingDocumentInfo',